complete CRUD functionality with authoriazation

+++ backend/controllers/goalcontroller.js
@@ -38,7 +38,12 @@ const updateGoal = expressHandler(async (req, res) => {
     res.status(401);
     throw new Error("User not found");
   }
-  
+
+  if(goal.user.toString() !== user.id){
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -54,6 +59,17 @@ const deleteGoal = expressHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+    const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {  
+    res.status(401);
+    throw new Error("User not found");
+  }
+  
+  if(goal.user.toString() !== user.id){
+    res.status(401);
+    throw new Error("User not authorized");
   }
   await Goal.findByIdAndDelete(req.params.id);
   res.status(200).json(`message: Deleted goal with id ${req.params.id}`);
